refactor: reorganize project structure for MVVM pattern alignment
