+++ lib/core/consts.dart
@@ -7,5 +7,9 @@ class AppConsts{
 class AppAPIs{
   const AppAPIs._();
 
-  static const String pokemon = 'https://pokeapi.co/api/v2/pokemon/';
+  static const String pokemonList = 'https://pokeapi.co/api/v2/pokemon/';
+  static const String pokemonSpecies = 'https://pokeapi.co/api/v2/pokemon-species/';
+  static String pokemonCries(int id) => 'https://pokeapi.co/api/v2/pokemon/$id/';
+  static String pokemonSprites(int id) => 'https://pokeapi.co/api/v2/pokemon/$id/';
+  static const String pokemonGeneral = 'https://pokeapi.co/api/v2/';
 }
