+++ lib/views/user_page.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';
 import 'package:myinfogame/core/consts.dart';
 import 'package:myinfogame/core/json_file_parser.dart';
-import 'package:myinfogame/data/user.dart';
+import 'package:myinfogame/models/user.dart';
 
 class UserPage extends StatefulWidget {
   const UserPage({super.key});
